requirements.txt
openai>=1.0.0
edge-tts>=7.0.0
moviepy>=2.0.0
Pillow>=10.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
schedule>=1.2.0
